requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
Pillow==10.0.0
pytesseract==0.3.10
pdf2image==1.16.3
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
spacy==3.6.1
transformers==4.33.2
torch==2.0.1
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
